requirements.txt
asgiref==3.8.1
attrs==25.1.0
cffi==1.17.1
cryptography==44.0.1
Django==5.1.6
django-pytest==0.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.2.1
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
celery==5.3.6
redis==5.0.0
kombu==5.3.1

